pre methods; sync policy docs
Made "pre" methods for synchronized contextSwitcher getter/setter, for debugging.

Made startKernel accept the contextSwitcher as a parameter, to cut down on usage of synchronized contextSwitcher getter.

Decided on policies about when to synchronize vs when to just rely on a semaphore for visibility assurance, and documented those policies in some javadocs.

Next: make pre methods for curRun getter/setter; make isStopped synchronized (override and call the interface's super)