curRun "pre" methods; isStopped super default
Made "pre" methods for currentlyRunning getter/setter.

Allowed Stoppable.isStopped to be synchronized in UCS implementations by overriding it in UCS and calling the super default.

Tried to synchronize isStopped but it blocked the Kernel, maybe due to some kind of reentrancy issue? So decided to leave it unsynchronized.

Next: allow timerThread to stop curRun